Login Logic Created
Login function is created.\ncreateJwt and attachJwtToCookie function is created.\nToken db model is created.\nToken model saves the refresh token for verification.\nLogin function checks if the user informations are correct and if there is a refresh token or not. If there is a refresh token then attack it to cookie with access and refresh tokens. If not, create new token schema and jwt tokens for client.