Forgot and Reset Password Functions
Forgot password function sends an email which is provided by the user to confirm email address. When the link is clicked it will redirect to client page with params passwordToken and email. In the redirected page, client must send back these with password in body request. After verification check in server, response will be given as success or an error.