Price function is seperated as another file to be used multiple controllers.
Limit and skip function is also created seperatly to use multiple times.

Single order and order models are created with interfaces

Static function for stock decrease in order and decrease on cancel is added to the model.

Create order, get single order and get all single orders controllers are created.